fix: seperate adjustSubscription test cases from billingRunHelpers.test.ts
